requirements.txt
# Qwen2.5-1.5B Jailbreak Lab - CPU dependencies
# No vllm (GPU only), no bitsandbytes (quantization requires GPU)

torch>=2.1.0
transformers>=4.50.0
peft>=0.15.0
trl>=0.18.0
datasets>=3.0.0
accelerate>=1.5.0
pyyaml>=6.0
rich>=13.0.0
